refactor(sdk): one resolver for the two path forms users type
A user-typed path field accepts two conveniences: "%VAR%" environment
references and Windows shell virtual paths ("shell:Downloads", "::{CLSID}").
Each call site parsed them itself, with slightly different trimming and case
rules -- PathAvailability and FavoritePathResolver both carried their own copy
of the same two steps, and App/Services/FileExecutor a third.

UserPathResolver is the SDK home those rules now live in. Its two steps stay
separate methods on purpose: callers that must keep a token virtual use
IsVirtualPath and never resolve, because a shell folder's localized display
name and a packaged app's identity are only reachable through the virtual
form. That same predicate replaces each hand-rolled "::"/"shell:" test, and
is how a caller tells an unresolvable input apart from a resolved one.

Resolve keeps FavoritePathResolver's injectable resolver parameter -- the seam
its tests drive the COM-backed lookup through, which is what lets the shared
implementation be covered without the shell actually being reached.

PathAvailability and FavoritePathResolver now delegate, keeping only the
policies that are actually theirs: validating an entry rather than resolving
it, and Favorites' web-url branches.
