+++ PluginSdk/Helpers/PathAvailability.cs
@@ -6,8 +6,8 @@ public static class PathAvailability
 {
     public static bool IsAvailable(string? path)
     {
-        var expanded = Expand(path);
-        if (IsVirtualPath(expanded))
+        var expanded = UserPathResolver.Expand(path);
+        if (UserPathResolver.IsVirtualPath(expanded))
         {
             return ShellVirtualPathValidator.Exists(expanded);
         }
@@ -17,19 +17,12 @@ public static class PathAvailability
 
     public static bool IsFolderAvailable(string? path)
     {
-        var expanded = Expand(path);
-        if (IsVirtualPath(expanded))
+        var expanded = UserPathResolver.Expand(path);
+        if (UserPathResolver.IsVirtualPath(expanded))
         {
             return ShellVirtualPathValidator.Exists(expanded, requireFolder: true);
         }
 
         return Directory.Exists(expanded);
     }
-
-    private static string Expand(string? path) =>
-        string.IsNullOrWhiteSpace(path) ? string.Empty : Environment.ExpandEnvironmentVariables(path.Trim());
-
-    private static bool IsVirtualPath(string path) =>
-        path.StartsWith("::", StringComparison.Ordinal)
-        || path.StartsWith("shell:", StringComparison.OrdinalIgnoreCase);
 }
